[SOL-7919] Make a mess of a packaging
Trying to figure out how in the world this all should come together. It
seems that there is no definitive consensus over how to organise the
packages and projects and what tools to use.

Trying to follow one best-practices guide here, but even that doesn't
discuss using Docker with development.

We shall see how it all turns out.