Add development tools for VSCode
Storing these settings to create development environment from scratch
may be a good idea. These work with Visua Studio Code and are meant to be
used for Remote Containers.

Using docker continers like this is an alternative for creating
virtual environments locally. But in it's advantage provides a way to have
any Python version / operating system available.