Parse each contour to separate item in list
As requested in Slack #asi-sol-sasse, have a single line representing
one contoured geometry. This means that at one point-in-time there can
be several different geometries.

This may be a bit more difficult when running the same
scipt more than once. As old lines may need to be removed instead of
just overwriting them.

We have to keep this in mind when writing the code for the main app
that's using the ORM-class.